LayoutTests/imported/w3c/web-platform-tests/hr-time/basic.worker-expected.txt


PASS WorkerGlobalScope.performance.now() is a function 
PASS WorkerGlobalScope.performance.now() returns a positive number 
PASS WorkerGlobalScope.performance.now() difference is not negative 
